Support encoding/decoding `MessagePackTimestamp` using non-MessagePack encoders/decoders.
`MessagePackTimestamp` currently requires a `MessagePackEncoder`/`MessagePackDecoder` for encoding/decoding. Using a non-MessagePack encoder/decoder results in a fatal error.

A client may want to use a human-readable serialization format, such as JSON, in automated tests to make it easier to read and write the test. However, if a `MessagePackTimestamp` is present anywhere in the data structure to be serialized/deserialized, the test will crash.

To enable the above use case, we add fallback behavior for non-MessagePack encoders/decoders when encoding/decoding `MessagePackTimestamp`.